Rename the package accordingly Prevent internal ID leaks
