+++ src/main/java/uk/gov/cshr/civilservant/Application.java
@@ -1,0 +1,15 @@
+package uk.gov.cshr.civilservant;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Main Spring application configuration and entry point.
+ */
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
