+++ src/main/java/uk/gov/cshr/civilservant/api/Results.java
@@ -1,0 +1,17 @@
+package uk.gov.cshr.civilservant.api;
+
+import java.util.Collection;
+import java.util.List;
+
+public class Results<T> {
+
+    private Collection<T> results;
+
+    Results(Collection<T> results) {
+        this.results = results;
+    }
+
+    public Collection<T> getResults() {
+        return results;
+    }
+}
