+++ src/main/java/uk/gov/cshr/civilservant/config/SecurityConfig.java
@@ -1,0 +1,29 @@
+package uk.gov.cshr.civilservant.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+import org.springframework.security.oauth2.provider.token.RemoteTokenServices;
+import org.springframework.security.oauth2.provider.token.ResourceServerTokenServices;
+
+@Configuration
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+@EnableResourceServer
+@EnableWebSecurity
+public class SecurityConfig extends ResourceServerConfigurerAdapter {
+
+    @Bean
+    public ResourceServerTokenServices tokenService(@Value("${oauth.clientId}") String clientId,
+                                                    @Value("${oauth.clientSecret}") String clientSecret,
+                                                    @Value("${oauth.checkTokenEndpointUrl}") String checkTokenEndpointUrl) {
+        RemoteTokenServices tokenServices = new RemoteTokenServices();
+        tokenServices.setClientId(clientId);
+        tokenServices.setClientSecret(clientSecret);
+        tokenServices.setCheckTokenEndpointUrl(checkTokenEndpointUrl);
+        return tokenServices;
+    }
+}
