--- src/main/java/uk/gov/cslearning/civilservant/Application.java
@@ -1,15 +1,0 @@
-package uk.gov.cslearning.civilservant;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Main Spring application configuration and entry point.
- */
-@SpringBootApplication
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
