--- src/main/java/uk/gov/cslearning/civilservant/api/Results.java
@@ -1,17 +1,0 @@
-package uk.gov.cslearning.civilservant.api;
-
-import java.util.Collection;
-import java.util.List;
-
-public class Results<T> {
-
-    private Collection<T> results;
-
-    Results(Collection<T> results) {
-        this.results = results;
-    }
-
-    public Collection<T> getResults() {
-        return results;
-    }
-}
