--- src/main/java/uk/gov/cslearning/civilservant/domain/Grade.java
@@ -1,83 +1,0 @@
-package uk.gov.cslearning.civilservant.domain;
-
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-
-import javax.persistence.*;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static org.apache.commons.lang3.StringUtils.isNotEmpty;
-
-@Entity
-public class Grade {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String code;
-
-    @Column(nullable = false)
-    private String name;
-
-    protected Grade() {
-    }
-
-    public Grade(String code, String name) {
-        setCode(code);
-        setName(name);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        checkArgument(isNotEmpty(code));
-        this.code = code;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        checkArgument(isNotEmpty(name));
-        this.name = name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Grade grade = (Grade) o;
-
-        return new EqualsBuilder()
-                .append(code, grade.code)
-                .isEquals();
-    }
-
-    @Override
-    public int hashCode() {
-        return new HashCodeBuilder()
-                .append(code)
-                .toHashCode();
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .append("id", id)
-                .append("code", code)
-                .append("name", name)
-                .toString();
-    }
-}
