--- src/main/java/uk/gov/cslearning/civilservant/domain/Identity.java
@@ -1,66 +1,0 @@
-package uk.gov.cslearning.civilservant.domain;
-
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-
-import javax.persistence.*;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static org.apache.commons.lang3.StringUtils.isNotBlank;
-
-@Entity
-public class Identity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String uid;
-
-    protected Identity() {
-    }
-
-    public Identity(String uid) {
-        checkArgument(isNotBlank(uid));
-        this.uid = uid;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getUid() {
-        return uid;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-
-        if (o == null || getClass() != o.getClass()) return false;
-
-        Identity identity = (Identity) o;
-
-        return new EqualsBuilder()
-                .append(uid, identity.uid)
-                .isEquals();
-    }
-
-    @Override
-    public int hashCode() {
-        return new HashCodeBuilder()
-                .append(uid)
-                .toHashCode();
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .append("id", id)
-                .append("uid", uid)
-                .toString();
-    }
-}
-
