--- src/main/java/uk/gov/cslearning/civilservant/repository/CivilServantRepository.java
@@ -1,14 +1,0 @@
-package uk.gov.cslearning.civilservant.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-import uk.gov.cslearning.civilservant.domain.CivilServant;
-import uk.gov.cslearning.civilservant.domain.Identity;
-
-import java.util.Optional;
-
-@Repository
-public interface CivilServantRepository extends CrudRepository<CivilServant, Long> {
-
-    Optional<CivilServant> findByIdentity(Identity identity);
-}
