--- src/main/java/uk/gov/cslearning/civilservant/repository/OrganisationRepository.java
@@ -1,13 +1,0 @@
-package uk.gov.cslearning.civilservant.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-import uk.gov.cslearning.civilservant.domain.Organisation;
-
-import java.util.Optional;
-
-@Repository
-public interface OrganisationRepository extends CrudRepository<Organisation, Long> {
-
-    Optional<Organisation> findByCode(String code);
-}
