--- src/test/java/uk/gov/cslearning/civilservant/repository/DepartmentRepositoryTest.java
@@ -1,34 +1,0 @@
-package uk.gov.cslearning.civilservant.repository;
-
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.cslearning.civilservant.domain.Department;
-
-import java.util.Optional;
-
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class DepartmentRepositoryTest {
-
-    @Autowired
-    private DepartmentRepository departmentRepository;
-
-    @Test
-    public void shouldFindDepartmentByExistingCode() {
-        Optional<Department> department = departmentRepository.findByCode("co");
-        assertTrue(department.isPresent());
-    }
-
-    @Test
-    public void shouldNotFindDepartmentByUnrecognisedCode() {
-        Optional<Department> department = departmentRepository.findByCode("unknown");
-        assertFalse(department.isPresent());
-    }
-}
